Fixed issues in the multiwii board driver

+++ boards/multiwii.h
@@ -17,7 +17,7 @@ extern "C" {
 #define PWM_MIN MINCOMMAND
 #define PWM_MAX 2000
 
-void get_accelerometer(int16_t *x, int16_t *y, int16_t *z); 
+void get_accelerometer(float *x, float *y, float *z); 
 void get_gyroscope(float *x, float *y, float *z); 
 void get_magnetometer(int16_t *x, int16_t *y, int16_t *z); 
 void get_altitude(int16_t *alt);
